% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_band.R
\name{albums_by_bands}
\alias{albums_by_bands}
\title{Calculates an artists albums ranked by release year}
\usage{
albums_by_bands(band.var)
}
\arguments{
\item{band.var}{the name of the band}
}
\value{
the sorted list of albums and rankings
}
\description{
Calculates an artists albums ranked by release year
}
\examples{
albums_by_bands("The Strokes")
albums_by_bands("The Smashing Pumpkins")
}
