% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_bands.R
\name{band_album_comparison_chart}
\alias{band_album_comparison_chart}
\title{Compare the rating of two bands' albums over time on a line chart}
\usage{
band_album_comparison_chart(var.artist1, var.artist2)
}
\arguments{
\item{var.artist1}{the first artist to compare, displayed as red}

\item{var.artist2}{the second artist to compare, displayed as blue}
}
\description{
Compare the rating of two bands' albums over time on a line chart
}
\examples{
band_album_comparison_chart("Elliott Smith", "Heatmiser")
band_album_comparison_chart("Blur", "Oasis")
}
