% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_band.R
\name{band_album_count}
\alias{band_album_count}
\title{Calculate an artist's number of ranked albums}
\usage{
band_album_count(band.var)
}
\arguments{
\item{band.var}{the name of the band}
}
\value{
the number of ranked albums by a band
}
\description{
Calculate an artist's number of ranked albums
}
\examples{
band_album_count("Buckethead")
band_album_count("Frank Zappa")
}
