% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_band.R
\name{band_mean_rating}
\alias{band_mean_rating}
\title{Calculate an artist's average rating and prints it}
\usage{
band_mean_rating(band.var)
}
\arguments{
\item{band.var}{the name of the band}
}
\value{
the artist's average rating
}
\description{
Calculate an artist's average rating and prints it
}
\examples{
band_mean_rating("The Get Up Kids")
band_mean_rating("The Beatles")
}
