% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinyl.R
\name{missing_vinyl}
\alias{missing_vinyl}
\title{Find top-rated albums that the person does not own on vinyl}
\usage{
missing_vinyl()
}
\value{
albums ranked 9 or higher not on vinyl
}
\description{
Find top-rated albums that the person does not own on vinyl
}
