% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_one_albums.R
\name{number_one_album}
\alias{number_one_album}
\title{List the Number 1 album for each year in the indicated range}
\usage{
number_one_album(var.startyear, var.endyear)
}
\arguments{
\item{var.startyear}{the starting year of the range, inclusive}

\item{var.endyear}{the ending year of the range, inclusive}
}
\value{
the top albums for the selected range
}
\description{
List the Number 1 album for each year in the indicated range
}
\examples{
number_one_album(1994, 2007)
number_one_album(2007, 2024)
}
