% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{server}
\alias{server}
\title{This function is the main server "function" with all of the logic}
\usage{
server(input, output)
}
\arguments{
\item{input}{the user input to the server to be processed}

\item{output}{the corresponding output to be displayed}
}
\description{
This function is the main server "function" with all of the logic
}
\examples{
shinyApp(ui = ui, server = server)
}
