% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_year.R
\name{year_albums}
\alias{year_albums}
\title{List the top albums for a year, in ranking order}
\usage{
year_albums(year.var)
}
\arguments{
\item{year.var}{the year to filter to}
}
\value{
the sorted list of albums for a year
}
\description{
List the top albums for a year, in ranking order
}
\examples{
year_albums(1999)
year_albums(1979)
}
