% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinyl.R
\name{year_most_vinyl}
\alias{year_most_vinyl}
\title{Sort years by vinyl ownership, from most to least}
\usage{
year_most_vinyl()
}
\value{
a sorted list of years and the number of vinyls owned, from most to least
}
\description{
Sort years by vinyl ownership, from most to least
}
